chore: revert splitting tests to reduce changeset size
