--- codebuild_specs/run_cdk_tests.yml
@@ -1,24 +1,0 @@
-version: 0.2
-env:
-  shell: bash
-  variables:
-    CI: true
-    CODEBUILD: true
-    NODE_OPTIONS: --max-old-space-size=8096
-phases:
-  build:
-    commands:
-      - codebuild-breakpoint
-      - source ./shared-scripts.sh && _setupCDKTestsLinux
-      - source ./shared-scripts.sh && _runCDKTestsLinux
-  post_build:
-    commands:
-      - source ./shared-scripts.sh && _disableLocalRegistry
-      - source ./shared-scripts.sh && _unassumeTestAccountCredentials
-      - aws sts get-caller-identity
-      - source ./shared-scripts.sh && _scanArtifacts
-
-artifacts:
-  files:
-    - '**/*'
-  base-directory: $CODEBUILD_SRC_DIR/packages/amplify-graphql-api-construct-tests/amplify-e2e-reports
