--- packages/amplify-graphql-api-construct-tests/.gitignore
@@ -1,3 +1,0 @@
-lib
-amplify-e2e-reports
-reports
