--- packages/amplify-graphql-api-construct-tests/environment.js
@@ -1,4 +1,0 @@
-// Allow people to use `amplify-e2e-core/environment` as a runner.
-const environment = require('./lib/cli-test-environment');
-
-module.exports = environment;
