--- packages/amplify-graphql-api-construct-tests/package.json
@@ -1,114 +1,0 @@
-{
-  "name": "amplify-graphql-api-construct-tests",
-  "version": "0.0.1",
-  "description": "CDK Construct E2e test suite",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/aws-amplify/amplify-category-api.git",
-    "directory": "packages/amplify-e2e-tests"
-  },
-  "author": "Amazon Web Services",
-  "license": "Apache-2.0",
-  "main": "index.js",
-  "keywords": [
-    "graphql",
-    "appsync",
-    "aws"
-  ],
-  "private": true,
-  "scripts": {
-    "e2e": "jest --verbose --forceExit",
-    "build-tests": "tsc --build tsconfig.tests.json",
-    "clean-stale-test-buckets": "ts-node ./src/cleanup-stale-test-buckets.ts",
-    "clean-e2e-resources": "ts-node ./src/cleanup-e2e-resources.ts"
-  },
-  "dependencies": {
-    "@aws-sdk/client-amplify": "^3.405.0",
-    "@aws-sdk/client-cloudformation": "^3.405.0",
-    "@aws-sdk/client-codebuild": "^3.405.0",
-    "@aws-sdk/client-iam": "^3.405.0",
-    "@aws-sdk/client-organizations": "^3.405.0",
-    "@aws-sdk/client-s3": "^3.405.0",
-    "@aws-sdk/client-sts": "^3.405.0",
-    "dotenv": "^8.2.0",
-    "fs-extra": "^8.1.0",
-    "lodash": "^4.17.21",
-    "node-fetch": "^2.6.7",
-    "yargs": "^15.1.0",
-    "rimraf": "^3.0.0",
-    "node-pty": "beta",
-    "chalk": "^4.1.1",
-    "strip-ansi": "^6.0.0",
-    "retimer": "2.0.0",
-    "uuid": "^8.3.2",
-    "jest-environment-node": "^26.6.2"
-  },
-  "devDependencies": {
-    "@types/lodash": "^4.14.149",
-    "ts-node": "^8.10.2"
-  },
-  "jest": {
-    "verbose": false,
-    "preset": "ts-jest",
-    "testRunner": "amplify-graphql-api-construct-tests/runner",
-    "testEnvironment": "amplify-graphql-api-construct-tests/environment",
-    "transform": {
-      "^.+\\.tsx?$": "ts-jest"
-    },
-    "testURL": "http://localhost",
-    "testRegex": "(src/__tests__/.*|\\.(test|spec))\\.(ts|tsx|js)$",
-    "testPathIgnorePatterns": [
-      "/node_modules/",
-      "lib"
-    ],
-    "collectCoverage": false,
-    "collectCoverageFrom": [
-      "src/**/*.ts",
-      "!**/node_modules/**",
-      "!src/__tests__/**",
-      "!**/*.d.ts"
-    ],
-    "reporters": [
-      "default",
-      "jest-junit",
-      [
-        "amplify-graphql-api-construct-tests/reporter",
-        {
-          "publicPath": "./amplify-e2e-reports",
-          "filename": "index.html",
-          "expand": true
-        }
-      ],
-      [
-        "amplify-graphql-api-construct-tests/failed-test-reporter",
-        {
-          "publicPath": "./amplify-e2e-reports",
-          "reportPath": "./amplify-e2e-reports/amplify-e2e-failed-test.txt"
-        }
-      ]
-    ],
-    "moduleFileExtensions": [
-      "ts",
-      "tsx",
-      "js",
-      "jsx",
-      "json",
-      "node"
-    ],
-    "setupFilesAfterEnv": [
-      "<rootDir>/src/setup-tests.ts"
-    ],
-    "globals": {
-      "window": {},
-      "ts-jest": {
-        "diagnostics": false
-      }
-    }
-  },
-  "jest-junit": {
-    "outputDirectory": "reports/junit/",
-    "outputName": "js-test-results.xml",
-    "usePathForSuiteName": "true",
-    "addFileAttribute": "true"
-  }
-}
