--- packages/amplify-graphql-api-construct-tests/reporter.js
@@ -1,4 +1,0 @@
-// Allow people to use `amplify-graphql-api-construct-tests/environment` as a runner.
-const reporter = require('./lib/nexpect-reporter');
-
-module.exports = reporter;
