--- packages/amplify-graphql-api-construct-tests/runner.js
@@ -1,4 +1,0 @@
-// Allow people to use `amplify-graphql-api-construct-tests/runner` as a runner.
-const runner = require('./lib/cli-test-runner');
-
-module.exports = runner;
