--- packages/amplify-graphql-api-construct-tests/sample.env
@@ -1,25 +1,0 @@
-# Used for setting up a new profile
-AWS_ACCESS_KEY_ID=<your-access-key>
-AWS_SECRET_ACCESS_KEY=<your-secret-access-key>
-AWS_SESSION_TOKEN=<optional-session-token>
-
-# Used for Auth Hosted UI
-FACEBOOK_APP_ID=fbAppId
-FACEBOOK_APP_SECRET=fbAppSecret
-
-GOOGLE_APP_ID=gglAppID
-GOOGLE_APP_SECRET=gglAppSecret
-
-AMAZON_APP_ID=amaznAppID
-AMAZON_APP_SECRET=amaznAppSecret
-
-# the following keys are invalidated but they pass Cognito validation
-APPLE_APP_ID=com.fake.app
-APPLE_TEAM_ID=2QLEWNDK6K
-APPLE_KEY_ID=2QLZXKYJ8J
-APPLE_PRIVATE_KEY=MIGTAgEAMBMGByqGSM49AgEGCCqGSM49AwEHBHkwdwIBAQQgIltgNsTgTfSzUadYiCS0VYtDDMFln/J8i1yJsSIw5g+gCgYIKoZIzj0DAQehRANCAASI8E0L/DhR/mIfTT07v3VwQu6q8I76lgn7kFhT0HvWoLuHKGQFcFkXXCgztgBrprzd419mUChAnKE6y89bWcNw
-
-#Used for delete test
-AWS_ACCESS_KEY_ID=<your-access-key>
-AWS_SECRET_ACCESS_KEY=<your-secret-access-key>
-CLI_REGION=<region>
