--- packages/amplify-graphql-api-construct-tests/src/setup-tests.ts
@@ -1,4 +1,0 @@
-// tslint:disable-next-line: no-magic-numbers
-const JEST_TIMEOUT = 1000 * 60 * 60; // 1 hour
-
-jest.setTimeout(JEST_TIMEOUT);
