--- packages/amplify-graphql-api-construct-tests/src/sleep.ts
@@ -1,1 +1,0 @@
-export const sleep = async (milliseconds: number): Promise<void> => new Promise((resolve) => setTimeout(resolve, milliseconds));
