--- packages/amplify-graphql-api-construct-tests/tsconfig.tests.json
@@ -1,18 +1,0 @@
-{
-  "compilerOptions": {
-    "skipLibCheck": true,
-    "target": "es6",
-    "allowJs": true,
-    "module": "commonjs",
-    "sourceMap": true,
-    "esModuleInterop": true,
-    "rootDir": "./src",
-    "types": ["jest", "node"],
-    "outDir": "lib",
-    "lib": ["es2015", "es2016.array.include", "esnext.asynciterable", "dom"],
-    "declaration": true,
-    "typeRoots": ["../../node_modules/@types", "node_modules/@types", "./typings"]
-  },
-  "include": ["src/**/*"],
-  "exclude": ["node_modules", "lib", "__tests__", "custom-resources", "overrides"]
-}
