+++ shared-scripts.sh
@@ -181,17 +181,6 @@ function _installCLIFromLocalRegistry {
     npm list -g --depth=1 | grep -e '@aws-amplify/amplify-category-api' -e 'amplify-codegen'
     unsetNpmRegistryUrl
 }
-function _setupRegistryTimeouts {
-    echo "Setup Registry Timeouts"
-    # set longer timeout to avoid socket timeout error
-    npm config set fetch-retry-mintimeout 20000
-    npm config set fetch-retry-maxtimeout 120000
-}
-function _disableLocalRegistry {
-    echo "Stop verdaccio"
-    source codebuild_specs/scripts/local_publish_helpers.sh
-    unsetNpmRegistryUrl
-}
 function _loadTestAccountCredentials {
     echo ASSUMING PARENT TEST ACCOUNT credentials
     session_id=$((1 + $RANDOM % 10000))
@@ -210,16 +199,7 @@ function _setupE2ETestsLinux {
     echo "Setup E2E Tests Linux"
     loadCacheFromBuildJob
     loadCache verdaccio-cache $CODEBUILD_SRC_DIR/../verdaccio-cache
-    _installCLIFromLocalRegistry  
-    _loadTestAccountCredentials
-    _setShell
-}
-
-function _setupCDKTestsLinux {
-    echo "Setup CDK Tests Linux"
-    loadCacheFromBuildJob
-    loadCache verdaccio-cache $CODEBUILD_SRC_DIR/../verdaccio-cache
-    _setupRegistryTimeouts
+    _installCLIFromLocalRegistry
     cd packages/amplify-graphql-api-construct
     yarn package
     _loadTestAccountCredentials
@@ -229,11 +209,6 @@ function _setupCDKTestsLinux {
 function _runE2ETestsLinux {
     echo "RUN E2E Tests Linux"
     retry runE2eTest
-}
-
-function _runCDKTestsLinux {
-    echo "RUN CDK Tests Linux"
-    retry runCDKTest
 }
 
 function _runGqlE2ETests {
@@ -413,23 +388,6 @@ function runE2eTest {
     if [ -z "$FIRST_RUN" ] || [ "$FIRST_RUN" == "true" ]; then
         echo "using Amplify CLI version: "$(amplify --version)
         cd $(pwd)/packages/amplify-e2e-tests
-    fi
-
-    if [ -f  $FAILED_TEST_REGEX_FILE ]; then
-        # read the content of failed tests
-        failedTests=$(<$FAILED_TEST_REGEX_FILE)
-        yarn run e2e --maxWorkers=4 $TEST_SUITE -t "$failedTests"
-    else
-        yarn run e2e --maxWorkers=4 $TEST_SUITE
-    fi
-}
-
-function runCDKTest {
-    FAILED_TEST_REGEX_FILE="./amplify-e2e-reports/amplify-e2e-failed-test.txt"
-
-    if [ -z "$FIRST_RUN" ] || [ "$FIRST_RUN" == "true" ]; then
-        echo "using Amplify CLI version: "$(amplify --version)
-        cd $(pwd)/packages/amplify-graphql-api-construct-tests
     fi
 
     if [ -f  $FAILED_TEST_REGEX_FILE ]; then
